コメントの修正
