+++ app/src/main/java/com/example/myaku_rismu/feature/setting/SettingState.kt
@@ -11,3 +11,10 @@ data class SettingState(
     val dialog: SettingType? = null,
     val isHealthConnectLinked: Boolean = false,
 )
+
+data class InfoItemData(
+    val label: String,
+    val value: String,
+    val onClick: () -> Unit,
+    val isSelected: Boolean
+)
